\name{H0}
\alias{H0}
\title{
Cumulative Hazard Function
}
\description{
Cumulative Hazard Function for Exponential and Weibull Distributions
}
\usage{
H0(t, survdist, hazard, k)
}
\arguments{
  \item{t}{
}
  \item{survdist}{
survival distribution of uncured patients, it can be exponential or weibull
}
  \item{hazard}{
survival distribution of uncured patients, it can be exponential or weibull
}
  \item{k}{
the shape parameter of weibull distribution. By default, it is set to be 1 for exponential distribution
}
}
