\name{Sc}
\alias{Sc}
\title{
Sc Function
}
\description{
Survival distribution of censoring times
}
\usage{
Sc(t, accrualtime, followuptime, accrualdist)
}
\arguments{
  \item{t}{
time variable
}
  \item{accrualtime}{
length of accrual period.
}
  \item{followuptime}{
length of follow-up time.
}
  \item{accrualdist}{
accrual pattern. It can be "\code{uniform}", "\code{increasing}" or "\code{decreasing}".
}
}
