\name{e1684szdata}
\alias{e1684szdata}
\docType{data}
\title{
Eastern Cooperative Oncology Group (ECOG) data with treatment as only covariate
}
\description{
Sample Size Calculation Example Data
}
\usage{data(e1684szdata)}
\format{
  A data frame with 285 observations on the following 4 variables.
  \describe{
    \item{\code{Time}}{a numeric vector}
    \item{\code{Status}}{a numeric vector}
    \item{\code{X}}{a numeric vector}
    \item{\code{Z}}{a numeric vector}
  }
}

\examples{
data(e1684szdata)
## maybe str(e1684szdata) ; plot(e1684szdata) ...
}
\keyword{datasets}
