\name{NPHMC-package}
\alias{NPHMC-package}
\docType{package}
\title{
An R-package for Estimating Sample Size of Proportional Hazards Mixture Cure Model
}
\description{
Estimating sample size for survival trial with or without cure fractions
}
\details{
\tabular{ll}{
Package: \tab NPHMC\cr
Type: \tab Package\cr
Version: \tab 2.2\cr
Date: \tab 2013-09-23\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Chao Cai, Songfeng Wang, Wenbin Lu, Jiajia Zhang

Maintainer: Chao Cai <caic@email.sc.edu>
}


\references{
S. Wang, J. Zhang, and W. Lu. Sample size calculation for the proportional hazards cure model. Statistics in medicine, 31:3959-3971, 2012

C. Cai, et al., smcure: An R-Package for estimating semiparametric mixture cure models. Computer Methods and Programs in Biomedicine, 108(3):1255-60, 2012

}

\seealso{
\code{\link[smcure]{smcure}}
}
