\name{Sc}
\alias{Sc}
\title{
Sc function
}
\description{
Survival Distribution of Censoring times
}
\usage{
Sc(t, accrualtime, followuptime, accrualdist)
}
\arguments{
  \item{t}{
time variable
}
  \item{accrualtime}{
accrual time. The default is 3
}
  \item{followuptime}{
follow up time. The default is 4
}
  \item{accrualdist}{
accrual rate distribution. It can be "uniform", "increasing" or "decreasing"
}
}
